Update fibers and ios-sim usage
Update package.json to use fibers version 1.0.5.0 (it is the same as the currently used from branch vladimirov/fibers_1_0_5_fixes.
Use ios-sim-portable version 1.0.1 - it is the same as the master branch.

IMPORTANT: This pull request can be merged ONLY AFTER we publish version 1.0.1 of ios-sim-portable.  This requires merge of https://github.com/telerik/ios-sim-portable/pull/16